+++ koda_validate/dictionary.py
@@ -18,7 +18,26 @@ from typing import (
 
 from koda import Just, Maybe, mapping_get, nothing
 
-from koda_validate._generics import T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, A, Ret
+from koda_validate._generics import (
+    T1,
+    T2,
+    T3,
+    T4,
+    T5,
+    T6,
+    T7,
+    T8,
+    T9,
+    T10,
+    T11,
+    T12,
+    T13,
+    T14,
+    T15,
+    T16,
+    A,
+    Ret,
+)
 from koda_validate._internals import OBJECT_ERRORS_FIELD, _async_predicates_warning
 from koda_validate.base import (
     Predicate,
@@ -38,7 +57,10 @@ EXPECTED_MAP_ERR: Final[Invalid[Serializable]] = Invalid(
     {OBJECT_ERRORS_FIELD: ["expected a map"]}
 )
 
+VALID_NOTHING: Final[Validated[Maybe[Any], Any]] = Valid(nothing)
+
 KEY_MISSING_MSG: Final[Serializable] = ["key missing"]
+KEY_MISSING_ERR: Final[Invalid[Serializable]] = Invalid(KEY_MISSING_MSG)
 
 
 class KeyNotRequired(Validator[Any, Maybe[A], Serializable]):
@@ -491,6 +513,183 @@ class RecordValidator(_ToTupleValidatorUnsafe[Any, Ret, Serializable]):
     ) -> None:
         ...  # pragma: no cover
 
+    @overload
+    def __init__(
+        self,
+        *,
+        into: Callable[[T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11], Ret],
+        keys: Tuple[
+            KeyValidator[T1],
+            KeyValidator[T2],
+            KeyValidator[T3],
+            KeyValidator[T4],
+            KeyValidator[T5],
+            KeyValidator[T6],
+            KeyValidator[T7],
+            KeyValidator[T8],
+            KeyValidator[T9],
+            KeyValidator[T10],
+            KeyValidator[T11],
+        ],
+        validate_object: Optional[Callable[[Ret], Validated[Ret, Serializable]]] = None,
+        validate_object_async: Optional[
+            Callable[[Ret], Awaitable[Validated[Ret, Serializable]]]
+        ] = None,
+        preprocessors: Optional[List[Processor[Dict[Any, Any]]]] = None,
+    ) -> None:
+        ...  # pragma: no cover
+
+    @overload
+    def __init__(
+        self,
+        *,
+        into: Callable[[T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12], Ret],
+        keys: Tuple[
+            KeyValidator[T1],
+            KeyValidator[T2],
+            KeyValidator[T3],
+            KeyValidator[T4],
+            KeyValidator[T5],
+            KeyValidator[T6],
+            KeyValidator[T7],
+            KeyValidator[T8],
+            KeyValidator[T9],
+            KeyValidator[T10],
+            KeyValidator[T11],
+            KeyValidator[T12],
+        ],
+        validate_object: Optional[Callable[[Ret], Validated[Ret, Serializable]]] = None,
+        validate_object_async: Optional[
+            Callable[[Ret], Awaitable[Validated[Ret, Serializable]]]
+        ] = None,
+        preprocessors: Optional[List[Processor[Dict[Any, Any]]]] = None,
+    ) -> None:
+        ...  # pragma: no cover
+
+    @overload
+    def __init__(
+        self,
+        *,
+        into: Callable[[T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13], Ret],
+        keys: Tuple[
+            KeyValidator[T1],
+            KeyValidator[T2],
+            KeyValidator[T3],
+            KeyValidator[T4],
+            KeyValidator[T5],
+            KeyValidator[T6],
+            KeyValidator[T7],
+            KeyValidator[T8],
+            KeyValidator[T9],
+            KeyValidator[T10],
+            KeyValidator[T11],
+            KeyValidator[T12],
+            KeyValidator[T13],
+        ],
+        validate_object: Optional[Callable[[Ret], Validated[Ret, Serializable]]] = None,
+        validate_object_async: Optional[
+            Callable[[Ret], Awaitable[Validated[Ret, Serializable]]]
+        ] = None,
+        preprocessors: Optional[List[Processor[Dict[Any, Any]]]] = None,
+    ) -> None:
+        ...  # pragma: no cover
+
+    @overload
+    def __init__(
+        self,
+        *,
+        into: Callable[
+            [T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14], Ret
+        ],
+        keys: Tuple[
+            KeyValidator[T1],
+            KeyValidator[T2],
+            KeyValidator[T3],
+            KeyValidator[T4],
+            KeyValidator[T5],
+            KeyValidator[T6],
+            KeyValidator[T7],
+            KeyValidator[T8],
+            KeyValidator[T9],
+            KeyValidator[T10],
+            KeyValidator[T11],
+            KeyValidator[T12],
+            KeyValidator[T13],
+            KeyValidator[T14],
+        ],
+        validate_object: Optional[Callable[[Ret], Validated[Ret, Serializable]]] = None,
+        validate_object_async: Optional[
+            Callable[[Ret], Awaitable[Validated[Ret, Serializable]]]
+        ] = None,
+        preprocessors: Optional[List[Processor[Dict[Any, Any]]]] = None,
+    ) -> None:
+        ...  # pragma: no cover
+
+    @overload
+    def __init__(
+        self,
+        *,
+        into: Callable[
+            [T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15], Ret
+        ],
+        keys: Tuple[
+            KeyValidator[T1],
+            KeyValidator[T2],
+            KeyValidator[T3],
+            KeyValidator[T4],
+            KeyValidator[T5],
+            KeyValidator[T6],
+            KeyValidator[T7],
+            KeyValidator[T8],
+            KeyValidator[T9],
+            KeyValidator[T10],
+            KeyValidator[T11],
+            KeyValidator[T12],
+            KeyValidator[T13],
+            KeyValidator[T14],
+            KeyValidator[T15],
+        ],
+        validate_object: Optional[Callable[[Ret], Validated[Ret, Serializable]]] = None,
+        validate_object_async: Optional[
+            Callable[[Ret], Awaitable[Validated[Ret, Serializable]]]
+        ] = None,
+        preprocessors: Optional[List[Processor[Dict[Any, Any]]]] = None,
+    ) -> None:
+        ...  # pragma: no cover
+
+    @overload
+    def __init__(
+        self,
+        *,
+        into: Callable[
+            [T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16], Ret
+        ],
+        keys: Tuple[
+            KeyValidator[T1],
+            KeyValidator[T2],
+            KeyValidator[T3],
+            KeyValidator[T4],
+            KeyValidator[T5],
+            KeyValidator[T6],
+            KeyValidator[T7],
+            KeyValidator[T8],
+            KeyValidator[T9],
+            KeyValidator[T10],
+            KeyValidator[T11],
+            KeyValidator[T12],
+            KeyValidator[T13],
+            KeyValidator[T14],
+            KeyValidator[T15],
+            KeyValidator[T16],
+        ],
+        validate_object: Optional[Callable[[Ret], Validated[Ret, Serializable]]] = None,
+        validate_object_async: Optional[
+            Callable[[Ret], Awaitable[Validated[Ret, Serializable]]]
+        ] = None,
+        preprocessors: Optional[List[Processor[Dict[Any, Any]]]] = None,
+    ) -> None:
+        ...  # pragma: no cover
+
     def __init__(
         self,
         into: Union[
@@ -504,6 +703,17 @@ class RecordValidator(_ToTupleValidatorUnsafe[Any, Ret, Serializable]):
             Callable[[T1, T2, T3, T4, T5, T6, T7, T8], Ret],
             Callable[[T1, T2, T3, T4, T5, T6, T7, T8, T9], Ret],
             Callable[[T1, T2, T3, T4, T5, T6, T7, T8, T9, T10], Ret],
+            Callable[[T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11], Ret],
+            Callable[[T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12], Ret],
+            Callable[[T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13], Ret],
+            Callable[[T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14], Ret],
+            Callable[
+                [T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15], Ret
+            ],
+            Callable[
+                [T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15, T16],
+                Ret,
+            ],
         ],
         keys: Union[
             Tuple[KeyValidator[T1]],
@@ -567,6 +777,99 @@ class RecordValidator(_ToTupleValidatorUnsafe[Any, Ret, Serializable]):
                 KeyValidator[T9],
                 KeyValidator[T10],
             ],
+            Tuple[
+                KeyValidator[T1],
+                KeyValidator[T2],
+                KeyValidator[T3],
+                KeyValidator[T4],
+                KeyValidator[T5],
+                KeyValidator[T6],
+                KeyValidator[T7],
+                KeyValidator[T8],
+                KeyValidator[T9],
+                KeyValidator[T10],
+                KeyValidator[T11],
+            ],
+            Tuple[
+                KeyValidator[T1],
+                KeyValidator[T2],
+                KeyValidator[T3],
+                KeyValidator[T4],
+                KeyValidator[T5],
+                KeyValidator[T6],
+                KeyValidator[T7],
+                KeyValidator[T8],
+                KeyValidator[T9],
+                KeyValidator[T10],
+                KeyValidator[T11],
+                KeyValidator[T12],
+            ],
+            Tuple[
+                KeyValidator[T1],
+                KeyValidator[T2],
+                KeyValidator[T3],
+                KeyValidator[T4],
+                KeyValidator[T5],
+                KeyValidator[T6],
+                KeyValidator[T7],
+                KeyValidator[T8],
+                KeyValidator[T9],
+                KeyValidator[T10],
+                KeyValidator[T11],
+                KeyValidator[T12],
+                KeyValidator[T13],
+            ],
+            Tuple[
+                KeyValidator[T1],
+                KeyValidator[T2],
+                KeyValidator[T3],
+                KeyValidator[T4],
+                KeyValidator[T5],
+                KeyValidator[T6],
+                KeyValidator[T7],
+                KeyValidator[T8],
+                KeyValidator[T9],
+                KeyValidator[T10],
+                KeyValidator[T11],
+                KeyValidator[T12],
+                KeyValidator[T13],
+                KeyValidator[T14],
+            ],
+            Tuple[
+                KeyValidator[T1],
+                KeyValidator[T2],
+                KeyValidator[T3],
+                KeyValidator[T4],
+                KeyValidator[T5],
+                KeyValidator[T6],
+                KeyValidator[T7],
+                KeyValidator[T8],
+                KeyValidator[T9],
+                KeyValidator[T10],
+                KeyValidator[T11],
+                KeyValidator[T12],
+                KeyValidator[T13],
+                KeyValidator[T14],
+                KeyValidator[T15],
+            ],
+            Tuple[
+                KeyValidator[T1],
+                KeyValidator[T2],
+                KeyValidator[T3],
+                KeyValidator[T4],
+                KeyValidator[T5],
+                KeyValidator[T6],
+                KeyValidator[T7],
+                KeyValidator[T8],
+                KeyValidator[T9],
+                KeyValidator[T10],
+                KeyValidator[T11],
+                KeyValidator[T12],
+                KeyValidator[T13],
+                KeyValidator[T14],
+                KeyValidator[T15],
+                KeyValidator[T16],
+            ],
         ],
         validate_object: Optional[Callable[[Ret], Validated[Ret, Serializable]]] = None,
         validate_object_async: Optional[
@@ -606,7 +909,7 @@ class RecordValidator(_ToTupleValidatorUnsafe[Any, Ret, Serializable]):
         if not isinstance(data, dict):
             return EXPECTED_DICT_ERR
 
-        if self.preprocessors is not None:
+        if self.preprocessors:
             for preproc in self.preprocessors:
                 data = preproc(data)
 
@@ -661,7 +964,7 @@ class RecordValidator(_ToTupleValidatorUnsafe[Any, Ret, Serializable]):
         if not isinstance(data, dict):
             return EXPECTED_DICT_ERR
 
-        if self.preprocessors is not None:
+        if self.preprocessors:
             for preproc in self.preprocessors:
                 data = preproc(data)
 
@@ -793,7 +1096,7 @@ class DictValidatorAny(_ToTupleValidatorUnsafe[Any, Any, Serializable]):
         if not isinstance(data, dict):
             return EXPECTED_DICT_ERR
 
-        if self.preprocessors is not None:
+        if self.preprocessors:
             for preproc in self.preprocessors:
                 data = preproc(data)
 
@@ -846,7 +1149,7 @@ class DictValidatorAny(_ToTupleValidatorUnsafe[Any, Any, Serializable]):
         if not isinstance(data, dict):
             return EXPECTED_DICT_ERR
 
-        if self.preprocessors is not None:
+        if self.preprocessors:
             for preproc in self.preprocessors:
                 data = preproc(data)